Clear /dev/shm before process startup to prevent crashloops (#364)
When vLLM crashes there is no guarantee that /dev/shm will
be cleaned up. Since we are the only consumer of the volume
ensure it is empty before the process starts. Otherwise we
may run out of shm for NCCL init during subsequent crashes
converting temporary crashes to a permanent one.

Fixes issue #352

Signed-off-by: Clayton Coleman <smarterclayton@gmail.com>