Update GKE to align to UBI images (#415)
* Remove the GKE image now that the NVSHMEM issue is bypassed

Signed-off-by: Clayton Coleman <smarterclayton@gmail.com>

* Only 2Gi /dev/shm is necessary for NCCL on wide EP

Signed-off-by: Clayton Coleman <smarterclayton@gmail.com>

* Update GKE configuration for WideEP to use upstream image

A new variant `gke-a4` is necessary for B200 as the default memory
configuration results in an OOM. This is an upstream bug on vLLM.

Signed-off-by: Clayton Coleman <smarterclayton@gmail.com>

* Use a lower default value for DP chunk size

DeepEP uses a default 1024 QP depth and asserts that QP depth is
greater than or equal to `2 * (chunk_size + 1)`. Increasing
default QP depth increases static NVSHMEM allocation and may also
require further tuning.

Signed-off-by: Clayton Coleman <smarterclayton@gmail.com>

* Simplify startup probe and wait for 30m

Use the port name to avoid changing constants in multiple places.

Signed-off-by: Clayton Coleman <smarterclayton@gmail.com>

---------

Signed-off-by: Clayton Coleman <smarterclayton@gmail.com>